ai_evaluation/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
boto3==1.34.34
python-dotenv==1.0.0
